# api-contracts.yaml
openapi: 3.0.3
info:
  title: B2B Platform API
  version: 1.0.0
  description: >
    B2B procurement platform API: requests, keys, suppliers, moderation, email sending,
    inbound reply tracking. Attachments are uploaded separately and referenced by ID.
servers:
  - url: http://localhost:8000/api/v1
    description: Local
  - url: https://{host}/api/v1
    description: Production
    variables:
      host:
        default: app.example.com
...
security:
  - bearerAuth: []
paths:
...
  /user/upload-and-create:
    post:
      tags: [UserRequests]
      summary: Upload procurement file and create a draft request
      description: >
        Accepts PDF/DOCX/XLSX/TXT. For PDF: only up to PDF_MAX_PAGES pages are processed via LLM.
        DOCX/XLSX parsed with free parsers (no tokens).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Draft request created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreateRequestResponse" }
        "413":
          description: File too large / too many pages
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /user/requests:
    get:
      tags: [UserRequests]
      summary: List user requests
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
        - in: query
          name: offset
          schema: { type: integer, default: 0, minimum: 0 }
      responses:
        "200":
          description: List of requests
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RequestListResponse" }

    post:
      tags: [UserRequests]
      summary: Create a draft request from manual keys (no file)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateRequestManualRequest" }
      responses:
        "200":
          description: Draft request created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreateRequestResponse" }
  /user/requests/{requestId}:
    get:
      tags: [UserRequests]
      summary: Get request detail (keys + suppliers + statuses)
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Request detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RequestDetail" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /user/requests/{requestId}/keys:
    put:
      tags: [UserRequests]
      summary: Update/confirm keys for a draft request (after parsing)
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateRequestKeysRequest" }
      responses:
        "200":
          description: Updated request keys
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RequestDetail" }
  /user/requests/{requestId}/submit:
    post:
      tags: [UserRequests]
      summary: Submit request (start matching + create moderation tasks)
      description: >
        Matches suppliers by keys (exact + pg_trgm; embeddings later) and creates moderation tasks.
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubmitRequestResponse" }
        "400":
          description: Invalid state (e.g., not draft)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # ----------------------------
  # SUPPLIERS
  # ----------------------------
  /suppliers/search:
    get:
      tags: [Suppliers]
      summary: Search suppliers by query
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string, minLength: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SupplierSearchResponse" }

  # ----------------------------
  # ATTACHMENTS (correct way)
  # ----------------------------
...
  /user/requests/{requestId}/recipients:
    put:
      tags: [UserMessaging]
      summary: Set selected recipients (checkboxes) for the request
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateRecipientsRequest" }
      responses:
        "200":
          description: Updated recipients
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RecipientsResponse" }
  /user/requests/{requestId}/send:
    post:
      tags: [UserMessaging]
      summary: Send inquiry emails to selected suppliers
      description: >
        Sends one email per supplier per request (unique requestId+supplierId).
        Attachments are passed as attachment_ids.
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SendMessageRequestV2" }
      responses:
        "200":
          description: Send started / completed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SendMessageResponse" }
        "409":
          description: Already sent to some suppliers
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /user/requests/{requestId}/send-new:
    post:
      tags: [UserMessaging]
      summary: Send only to new suppliers (yellow status)
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SendMessageRequestV2" }
      responses:
        "200":
          description: Send started / completed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SendMessageResponse" }
  /user/requests/{requestId}/messages:
    get:
      tags: [UserMessaging]
      summary: List messages (thread) for the request (only the user can see)
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - in: query
          name: limit
          schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
        - in: query
          name: offset
          schema: { type: integer, default: 0, minimum: 0 }
        - in: query
          name: include_deleted
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageListResponse" }
...
  /user/blacklist/inn:
    post:
      tags: [Blacklist]
      summary: Add company INN to user's personal blacklist (user-level)
      description: >
        Personal blacklist. Prevent sending to this company for this user.
        UI may display supplier_name and checko_data when available.
        When added, system must auto-unselect related recipients (selected=false) for this user's requests.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddUserBlacklistInnRequest" }
      responses:
        "200":
          description: Added
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GenericOkResponse" }
    get:
      tags: [Blacklist]
      summary: List user's personal blacklist (by INN)
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 200, minimum: 1, maximum: 500 }
      responses:
        "200":
          description: Blacklist
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserBlacklistInnListResponse" }
  /user/blacklist/inn/{inn}:
    delete:
      tags: [Blacklist]
      summary: Remove INN from user's personal blacklist
      parameters:
        - in: path
          name: inn
          required: true
          schema: { type: string, minLength: 10, maxLength: 12 }
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GenericOkResponse" }

    delete:
      tags: [UserMessaging]
      summary: Delete message (soft delete) if email_policy=allow_delete
      description: >
        In append_only mode returns 403.
        Deletion is soft (hidden for the user), audit stays.
      parameters:
        - $ref: "#/components/parameters/MessageId"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GenericOkResponse" }
        "403":
          description: Not allowed by policy
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # ----------------------------
  # MODERATOR TASKS
  # ----------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RequestId:
      in: path
      name: requestId
      required: true
      schema: { type: integer, minimum: 1 }
...
    MessageId:
      in: path
      name: messageId
      required: true
      schema: { type: integer, minimum: 1 }
...