# backend/app/transport/routers/user_blacklist_inn.py
from fastapi import APIRouter, Depends, Header, HTTPException
from sqlalchemy.ext.asyncio import AsyncSession

from app.adapters.db.session import get_db_session
from app.adapters.db.repositories import UserBlacklistInnRepository
from app.transport.schemas.blacklist import (
    AddUserBlacklistInnRequestDTO,
    UserBlacklistInnItemDTO,
    UserBlacklistInnListResponseDTO,
)
from app.usecases.add_user_blacklist_inn import AddUserBlacklistInnUseCase
from app.usecases.list_user_blacklist_inn import ListUserBlacklistInnUseCase
from app.usecases.remove_user_blacklist_inn import RemoveUserBlacklistInnUseCase

router = APIRouter(tags=["Blacklist"])


def _require_user_id(authorization: str | None) -> int:
    # Temporary: aligns with current dev auth stub in /auth/me
    if not authorization or not authorization.startswith("Bearer "):
        raise HTTPException(status_code=401, detail="Missing Bearer token")
    return 1


@router.post("/user/blacklist/inn")
async def add_user_blacklist_inn(
    payload: AddUserBlacklistInnRequestDTO,
    authorization: str | None = Header(default=None),
    db: AsyncSession = Depends(get_db_session),
):
    user_id = _require_user_id(authorization)
    repo = UserBlacklistInnRepository(db)
    await AddUserBlacklistInnUseCase(repo).execute(user_id=user_id, inn=payload.inn, reason=payload.reason)
    return {"success": True}


@router.get("/user/blacklist/inn", response_model=UserBlacklistInnListResponseDTO)
async def list_user_blacklist_inn(
    limit: int = 200,
    authorization: str | None = Header(default=None),
    db: AsyncSession = Depends(get_db_session),
):
    user_id = _require_user_id(authorization)
    repo = UserBlacklistInnRepository(db)
    items = await ListUserBlacklistInnUseCase(repo).execute(user_id=user_id, limit=limit)
    return {"items": [UserBlacklistInnItemDTO(**x) for x in items]}


@router.delete("/user/blacklist/inn/{inn}")
async def remove_user_blacklist_inn(
    inn: str,
    authorization: str | None = Header(default=None),
    db: AsyncSession = Depends(get_db_session),
):
    user_id = _require_user_id(authorization)
    repo = UserBlacklistInnRepository(db)
    await RemoveUserBlacklistInnUseCase(repo).execute(user_id=user_id, inn=inn)
    return {"success": True}